// packages/JnEditor/interface/tinymce.d.ts
declare type EditorEvent<T> = NormalizedEvent<T>
// ...
interface RawEditorSettings extends BaseEditorSettings {
    external_plugins?: Record<string, string>
    mobile?: RawEditorSettings
    plugins?: string | string[]
}
// ...
interface Rect {
    inflate: (rect: GeomRect, w: number, h: number) => GeomRect
    relativePosition: (rect: GeomRect, targetRect: GeomRect, rel: string) => GeomRect
    findBestRelativePosition: (
        rect: GeomRect,
        targetRect: GeomRect,
        constrainRect: GeomRect,
        rels: string[]
    ) => string | null
    intersect: (rect: GeomRect, cropRect: GeomRect) => GeomRect | null
    clamp: (rect: GeomRect, clampRect: GeomRect, fixedSize?: boolean) => GeomRect
    create: (x: number, y: number, w: number, h: number) => GeomRect
    fromClientRect: (clientRect: DOMRect) => GeomRect
}
// ...
interface URIConstructor {
    readonly prototype: URI
    new (url: string, settings?: URISettings): URI
    getDocumentBaseUrl: (loc: {
        protocol: string
        host?: string
        href?: string
        pathname?: string
    }) => string
    parseDataUri: (uri: string) => {
        type: string
        data: string
    }
}
// ...
interface EditorManager extends Observable<EditorManagerEventMap> {
    $: DomQueryConstructor
    defaultSettings: RawEditorSettings
    majorVersion: string
    minorVersion: string
    releaseDate: string
    editors: Editor[]
    activeEditor: Editor
    focusedEditor: Editor
    settings: RawEditorSettings
    baseURI: URI
    baseURL: string
    documentBaseURL: string
    i18n: I18n
    suffix: string
    add(this: EditorManager, editor: Editor): Editor
    addI18n: (code: string, item: Record<string, string>) => void
    createEditor(this: EditorManager, id: string, settings: RawEditorSettings): Editor
    execCommand(this: EditorManager, cmd: string, ui: boolean, value: any): boolean
    get(this: EditorManager): Editor[]
    get(this: EditorManager, id: number | string): Editor
    init(this: EditorManager, settings: RawEditorSettings): Promise<Editor[]>
    overrideDefaults(this: EditorManager, defaultSettings: Partial<RawEditorSettings>): void
    remove(this: EditorManager): void
    remove(this: EditorManager, selector: string | Editor): Editor | void
    setActive(this: EditorManager, editor: Editor): void
    setup(this: EditorManager): void
    translate: (text: Untranslated) => TranslatedString
    triggerSave: () => void
    _setBaseUrl(this: EditorManager, baseUrl: string): void
}
// ...
declare class Editor implements EditorObservable {
    documentBaseUrl: string
    baseUri: URI
    settings: EditorSettings
    id: string
    plugins: Record<string, Plugin>
    documentBaseURI: URI
    baseURI: URI
    contentCSS: string[]
    contentStyles: string[]
    ui: EditorUi
    mode: EditorMode
    setMode: (mode: string) => void
    $: DomQueryConstructor
    shortcuts: Shortcuts
    loadedCSS: Record<string, any>
    editorCommands: EditorCommands
    suffix: string
    editorManager: EditorManager
    inline: boolean
    isNotDirty: boolean
    callbackLookup: any
    _nodeChangeDispatcher: NodeChange
    editorUpload: EditorUpload
    annotator: Annotator
    bodyElement: HTMLElement
    bookmark: any
    composing: boolean
    container: HTMLElement
    contentAreaContainer: HTMLElement
    contentDocument: Document
    contentWindow: Window
    delegates: Record<string, (event: any) => void>
    destroyed: boolean
    dom: DOMUtils
    editorContainer: HTMLElement
    eventRoot?: Element
    formatter: Formatter
    formElement: HTMLElement
    formEventDelegate: (e: Event) => void
    hasHiddenInput: boolean
    hasVisual: boolean
    hidden: boolean
    iframeElement: HTMLIFrameElement | null
    iframeHTML: string
    initialized: boolean
    notificationManager: NotificationManager
    orgDisplay: string
    orgVisibility: string
    parser: DomParser
    quirks: Quirks
    readonly: boolean
    removed: boolean
    schema: Schema
    selection: EditorSelection
    serializer: DomSerializer
    startContent: string
    targetElm: HTMLElement
    theme: Theme
    undoManager: UndoManager
    validate: boolean
    windowManager: WindowManager
    _beforeUnload: () => void
    _eventDispatcher: EventDispatcher<NativeEventMap>
    _mceOldSubmit: any
    _pendingNativeEvents: string[]
    _selectionOverrides: SelectionOverrides
    _skinLoaded: boolean
    bindPendingEventDelegates: EditorObservable['bindPendingEventDelegates']
    toggleNativeEvent: EditorObservable['toggleNativeEvent']
    unbindAllNativeEvents: EditorObservable['unbindAllNativeEvents']
    fire: EditorObservable['fire']
    on: EditorObservable['on']
    off: EditorObservable['off']
    once: EditorObservable['once']
    hasEventListeners: EditorObservable['hasEventListeners']
    constructor(id: string, settings: RawEditorSettings, editorManager: EditorManager)
    render(): void
    focus(skipFocus?: boolean): void
    hasFocus(): boolean
    execCallback(name: string, ...x: any[]): any
    translate(text: Untranslated): TranslatedString
    getParam<K extends keyof ParamTypeMap>(
        name: string,
        defaultVal: ParamTypeMap[K],
        type: K
    ): ParamTypeMap[K]
    getParam<K extends keyof EditorSettings>(
        name: K,
        defaultVal?: EditorSettings[K],
        type?: string
    ): EditorSettings[K]
    getParam<T>(name: string, defaultVal: T, type?: string): T
    hasPlugin(name: string, loaded?: boolean): boolean
    nodeChanged(args?: any): void
    addCommand(name: string, callback: EditorCommandCallback, scope?: object): void
    addQueryStateHandler(name: string, callback: () => boolean, scope?: any): void
    addQueryValueHandler(name: string, callback: () => string, scope?: any): void
    addShortcut(
        pattern: string,
        desc: string,
        cmdFunc: string | [string, boolean, any] | (() => void),
        scope?: any
    ): void
    execCommand(cmd: string, ui?: boolean, value?: any, args?: any): boolean
    queryCommandState(cmd: string): boolean
    queryCommandValue(cmd: string): string
    queryCommandSupported(cmd: string): boolean
    show(): void
    hide(): void
    isHidden(): boolean
    setProgressState(state: boolean, time?: number): void
    load(args?: any): string
    save(args?: any): string
    setContent(content: string, args?: SetContentArgs): string
    setContent(content: AstNode, args?: SetContentArgs): AstNode
    setContent(content: Content, args?: SetContentArgs): Content
    getContent(
        args: {
            format: 'tree'
        } & GetContentArgs
    ): AstNode
    getContent(args?: GetContentArgs): string
    insertContent(content: string, args?: any): void
    resetContent(initialContent?: string): void
    isDirty(): boolean
    setDirty(state: boolean): void
    getContainer(): HTMLElement
    getContentAreaContainer(): HTMLElement
    getElement(): HTMLElement
    getWin(): Window
    getDoc(): Document
    getBody(): HTMLElement
    convertURL(url: string, name: string, elm?: any): string
    addVisual(elm?: HTMLElement): void
    remove(): void
    destroy(automatic?: boolean): void
    uploadImages(callback?: UploadCallback): Promise<UploadResult$1[]>
    _scanForImages(): Promise<BlobInfoImagePair[]>
    addButton(): void
    addSidebar(): void
    addMenuItem(): void
    addContextToolbar(): void
}
// ...
interface Class {
    extend<T extends Props<A>, A extends any[] = any[]>(props: T): ExtendedClassConstructor<T, A>
}
interface RGB {
    r: number
    g: number
    b: number
}
interface HSV {
    h: number
    s: number
    v: number
}
declare type ColorConstructor = new (value?: string | RGB | HSV) => Color
interface Color {
    toRgb: () => RGB
    toHsv: () => HSV
    toHex: () => string
    parse: (value: string | RGB | HSV) => Color
}
interface DebounceFunc<T extends (...args: any[]) => void> {
    (...args: Parameters<T>): void
    stop: () => void
}
interface Delay {
    requestAnimationFrame: (callback: () => void, element?: HTMLElement) => void
    setEditorInterval: (editor: Editor, callback: () => void, time?: number) => number
    setEditorTimeout: (editor: Editor, callback: () => void, time?: number) => number
    setInterval: (callback: () => void, time?: number) => number
    setTimeout: (callback: () => void, time?: number) => number
    clearInterval: (id?: number) => void
    clearTimeout: (id?: number) => void
    debounce: <T extends (...args: any[]) => any>(callback: T, time?: number) => DebounceFunc<T>
    throttle: <T extends (...args: any[]) => any>(callback: T, time?: number) => DebounceFunc<T>
}
// ...
interface VK {
    BACKSPACE: number
    DELETE: number
    DOWN: number
    ENTER: number
    ESC: number
    LEFT: number
    RIGHT: number
    SPACEBAR: number
    TAB: number
    UP: number
    PAGE_UP: number
    PAGE_DOWN: number
    END: number
    HOME: number
    modifierPressed: (e: KeyboardLikeEvent) => boolean
    metaKeyPressed: (e: KeyboardLikeEvent) => boolean
}
// ...
interface XHR extends Observable<XHREventMap> {
    send(this: XHR, settings: XHRSettings): void
}
// ...
interface TinyMCE extends EditorManager {
    geom: {
        Rect: Rect
    }
    util: {
        Promise: PromiseConstructor
        Delay: Delay
        Tools: Tools
        VK: VK
        URI: URIConstructor
        Class: Class
        EventDispatcher: EventDispatcherConstructor<any>
        Observable: Observable<any>
        I18n: I18n
        XHR: XHR
        JSON: JSONUtils
        JSONRequest: JSONRequestConstructor
        JSONP: JSONP
        LocalStorage: Storage
        Color: ColorConstructor
        ImageUploader: ImageUploader
    }
    dom: {
        EventUtils: EventUtilsConstructor
        Sizzle: any
        DomQuery: DomQueryConstructor
        TreeWalker: DomTreeWalkerConstructor
        TextSeeker: new (dom: DOMUtils, isBlockBoundary?: (node: Node) => boolean) => TextSeeker
        DOMUtils: DOMUtilsNamespace
        ScriptLoader: ScriptLoaderConstructor
        RangeUtils: RangeUtilsNamespace
        Serializer: new (settings: DomSerializerSettings, editor?: Editor) => DomSerializer
        ControlSelection: (selection: EditorSelection, editor: Editor) => ControlSelection
        BookmarkManager: BookmarkManagerNamespace
        Selection: new (
            dom: DOMUtils,
            win: Window,
            serializer: DomSerializer,
            editor: Editor
        ) => EditorSelection
        StyleSheetLoader: new (
            documentOrShadowRoot: Document | ShadowRoot,
            settings: StyleSheetLoaderSettings
        ) => StyleSheetLoader
        Event: EventUtils
    }
    html: {
        Styles: new (settings?: StylesSettings, schema?: Schema) => Styles
        Entities: Entities
        Node: AstNodeConstructor
        Schema: new (settings?: SchemaSettings) => Schema
        SaxParser: SaxParserNamespace
        DomParser: new (settings?: DomParserSettings, schema?: Schema) => DomParser
        Writer: new (settings?: WriterSettings) => Writer
        Serializer: new (settings?: HtmlSerializerSettings, schema?: Schema) => HtmlSerializer
    }
    AddOnManager: AddOnManagerNamespace
    Annotator: new (editor: Editor) => Annotator
    Editor: EditorConstructor
    EditorCommands: EditorCommandsConstructor
    EditorManager: EditorManager
    EditorObservable: EditorObservable
    Env: Env
    FocusManager: FocusManager
    Formatter: new (editor: Editor) => Formatter
    NotificationManager: new (editor: Editor) => NotificationManager
    Shortcuts: ShortcutsConstructor
    UndoManager: new (editor: Editor) => UndoManager
    WindowManager: new (editor: Editor) => WindowManager
    DOM: DOMUtils
    ScriptLoader: ScriptLoader
    PluginManager: PluginManager
    ThemeManager: ThemeManager
    IconManager: IconManager
    Resource: Resource
    trim: Tools['trim']
    isArray: Tools['isArray']
    is: Tools['is']
    toArray: Tools['toArray']
    makeMap: Tools['makeMap']
    each: Tools['each']
    map: Tools['map']
    grep: Tools['grep']
    inArray: Tools['inArray']
    extend: Tools['extend']
    create: Tools['create']
    walk: Tools['walk']
    createNS: Tools['createNS']
    resolve: Tools['resolve']
    explode: Tools['explode']
    _addCacheSuffix: Tools['_addCacheSuffix']
    isOpera: boolean
    isWebKit: boolean
    isIE: false | number
    isGecko: boolean
    isMac: boolean
}